-- Read comment --
This file is an early description of what our system should look like / behave. Idea was to follow this and develop a functional code based on the psudo code listed here. We did come a bit short here in terms of coding capabilities, therfor we have a bit different code (but which follows some of these principles).